src/devui/requirements.txt
# Dependencies for Agent Framework + DevUI (local web UI)
#
# NOTE: At the time of writing, these may be published as pre-release packages.
# Install with: python -m pip install --pre -r src/devui/requirements.txt

agent-framework
agent-framework-azure-ai
agent-framework-devui

azure-identity
python-dotenv
